+++ src/testObjects.ts
@@ -33,8 +33,8 @@ export const testBackground: IBackground = {
   animation: {
     preset: 'network' as 'network',
     count: '150',
-    backgroundColour: '#e5e5e5',
-    particleColour: '#ffffff',
+    backgroundColour: { r: 20, g: 20, b: 20 },
+    particleColour: { r: 255, g: 255, b: 255 },
     repel: true,
   },
 }