+++ src/utils/colour.test.ts
@@ -1,5 +1,10 @@
 import { RGBColor } from 'react-color'
-import { RGBColorToString } from 'utils/colour'
+import { testBackground } from 'testObjects'
+import {
+  getGradientString,
+  RGBColorToHex,
+  RGBColorToString,
+} from 'utils/colour'
 
 describe('RGBColor conversion', () => {
   it('should correctly convert RGBColor values to string representations', () => {
@@ -57,5 +62,64 @@ describe('RGBColor conversion', () => {
     }
     expected = 'rgba(100, 100, 100, 1)'
     expect(RGBColorToString(colour)).toEqual(expected)
-  })
+  }),
+    it('should correctly convert RGBColor values to hex representations', () => {
+      let colour: RGBColor = {
+        r: 50,
+        g: 168,
+        b: 82,
+        a: 1,
+      }
+      let expected = '#32a852'
+      expect(RGBColorToHex(colour)).toEqual(expected)
+
+      colour = {
+        r: 235,
+        g: 213,
+        b: 234,
+        a: undefined,
+      }
+      expected = '#ebd5ea'
+      expect(RGBColorToHex(colour)).toEqual(expected)
+
+      colour = {
+        r: 1,
+        g: 0,
+        b: 1,
+        a: undefined,
+      }
+      expected = '#010001'
+      expect(RGBColorToHex(colour)).toEqual(expected)
+    }),
+    it('should correctly construct gradient strings', () => {
+      let gradient = {
+        ...testBackground.gradient,
+        type: 'linear',
+      }
+      let expected =
+        'linear-gradient(0deg, rgba(255, 255, 255, 1) 0%, rgba(210, 210, 210, 1) 100%)'
+      expect(
+        getGradientString(
+          gradient.startColour,
+          gradient.endColour,
+          gradient.angle,
+          gradient.type
+        )
+      ).toEqual(expected)
+
+      gradient = {
+        ...gradient,
+        type: 'radial',
+      }
+      expected =
+        'radial-gradient(circle, rgba(255, 255, 255, 1) 0%, rgba(210, 210, 210, 1) 100%)'
+      expect(
+        getGradientString(
+          gradient.startColour,
+          gradient.endColour,
+          gradient.angle,
+          gradient.type
+        )
+      ).toEqual(expected)
+    })
 })